Fixed type conflicts between NSUInteger and unsigned int on Ubuntu 64-bit.
Patch by Kjell Oyvind Konstali


git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile@6144 7b3f36aa-e6db-49c1-8b4c-66eac1997e64

+++ Frameworks/LuceneKit/Source/Search/LCFieldCacheImpl.m
@@ -77,7 +77,7 @@
 	return NO;
 }
 
-- (unsigned) hash
+- (NSUInteger) hash
 {
 	return [field hash] ^ type ^ ((custom == nil) ? 0 : [custom hash]) ^ ((locale == nil) ? 0 : [locale hash]);
 }
